@@ -9,7 +9,7 @@
     "build": "ng build",
     "lint": "./node_modules/.bin/tslint -c tslint.json 'src/**/*.ts'",
     "start:server": "nodemon",
-    "start:client": "ng serve --proxy-config proxy.config.json",
+    "start:client": "ng build --aot --watch=true",
     "start:dev": "concurrently --kill-others-on-fail \"yarn start:server\" \"yarn start:client\"",
     "prestart:prod": "rm -rf dist && tsc",
     "start:prod": "node dist/main.js",
@@ -22,19 +22,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.6",
-    "@angular/common": "^6.1.6",
-    "@angular/compiler": "^6.1.6",
-    "@angular/core": "^6.1.6",
-    "@angular/forms": "^6.1.6",
-    "@angular/http": "^6.1.6",
-    "@angular/platform-browser": "^6.1.6",
-    "@angular/platform-browser-dynamic": "^6.1.6",
-    "@angular/router": "^6.1.6",
-    "@nestjs/common": "^5.3.5",
-    "@nestjs/core": "^5.3.4",
+    "@angular/animations": "^6.1.7",
+    "@angular/cdk": "^6.2.0",
+    "@angular/common": "^6.1.7",
+    "@angular/compiler": "^6.1.7",
+    "@angular/core": "^6.1.7",
+    "@angular/forms": "^6.1.7",
+    "@angular/http": "^6.1.7",
+    "@angular/material": "^6.4.7",
+    "@angular/platform-browser": "^6.1.7",
+    "@angular/platform-browser-dynamic": "^6.1.7",
+    "@angular/router": "^6.1.7",
+    "@nestjs/common": "^5.3.7",
+    "@nestjs/core": "^5.3.7",
     "@nestjs/swagger": "^2.5.1",
-    "@nestjs/typeorm": "^5.2.0",
+    "@nestjs/typeorm": "^5.2.1",
+    "angular-oauth2-oidc": "^4.0.3",
     "class-transformer": "^0.1.9",
     "class-validator": "^0.9.1",
     "core-js": "^2.5.4",
@@ -43,42 +46,43 @@
     "mongodb": "~3.0.11",
     "reflect-metadata": "^0.1.12",
     "rxjs": "^6.3.2",
+    "rxjs-compat": "^6.3.2",
     "typeorm": "^0.2.7",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.5",
-    "@angular/cli": "~6.1.5",
-    "@angular/compiler-cli": "^6.1.6",
-    "@angular/language-service": "^6.1.6",
-    "@nestjs/testing": "^5.3.1",
+    "@angular-devkit/build-angular": "~0.8.2",
+    "@angular/cli": "~6.2.2",
+    "@angular/compiler-cli": "^6.1.7",
+    "@angular/language-service": "^6.1.7",
+    "@nestjs/testing": "^5.3.7",
     "@types/express": "^4.0.39",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
-    "@types/jest": "^23.3.1",
+    "@types/jest": "^23.3.2",
     "@types/node": "^10.9.4",
-    "@types/supertest": "^2.0.4",
+    "@types/supertest": "^2.0.6",
     "codelyzer": "~4.4.4",
     "concurrently": "^4.0.1",
     "jasmine-core": "~3.2.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^23.5.0",
+    "jest": "^23.6.0",
     "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.3",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^1.3.1",
     "nodemon": "^1.18.4",
     "prettier": "^1.11.1",
-    "protractor": "~5.4.0",
-    "supertest": "^3.2.0",
+    "protractor": "~5.4.1",
+    "supertest": "^3.3.0",
     "ts-jest": "^23.1.4",
     "ts-loader": "^4.1.0",
     "ts-node": "^7.0.1",
-    "tsconfig-paths": "^3.1.1",
+    "tsconfig-paths": "^3.6.0",
     "tslint": "~5.11.0",
     "typescript": "~2.7.2",
-    "webpack": "^4.17.2",
+    "webpack": "^4.19.0",
     "webpack-cli": "^3.1.0",
     "webpack-node-externals": "^1.6.0"
   },
