@@ -1,8 +1,14 @@
 import { Injectable } from '@nestjs/common';
+import { SetupService } from './controllers/setup/setup.service';
+import { PLEASE_RUN_SETUP } from './constants/messages';
 
 @Injectable()
 export class AppService {
-  message() {
-    return { message: 'NestJS' };
+  constructor(private readonly idpSetupService: SetupService) {}
+  async info() {
+    const info = (await this.idpSetupService.getInfo()) || {
+      message: PLEASE_RUN_SETUP,
+    };
+    return info;
   }
 }
